Scaffold API route package
